requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
 flask_login==0.6.3
pytest==8.4.2
python-dotenv==1.0.1
Flask-WTF==1.2.1
email-validator==2.1.0.post1
sentence-transformers
scikit-learn
numpy
flask-limiter>=4.0.0
